// FrontEnd/src/app/models/grafico-funil/GraficoFunil.ts
export class GraficoFunil {
	CodMarca: number;
	DescMarca: string;
	CorSiteMarca: string;
	PeriodoAnt: string;
	BaseAbsAtual: number;
	BaseAbsAnterior: number;
	ConhecimentoAnterior: number;
	ConhecimentoAtual: number;
	ConhecimentoTesteAtual: string;
	ConhecimentoADV: number;
	ConsideracaoAnterior: number;
	ConsideracaoAtual: number;
	ConsideracaoTesteAtual: string;
	ConsideracaoADV: number;
	UsoAnterior: number;
	UsoAtual: number;
	UsoTesteAtual: string;
	UsoADV: number;
	PreferenciaAnterior: number;
	PreferenciaAtual: number;
	PreferenciaTesteAtual: string;
	PreferenciaADV: number;

	LoyaltyAnterior: number;
	LoyaltyAtual: number;
	LoyaltyTesteAtual: string;
	LoyaltyADV: number;

	// 🔥 NOVOS CAMPOS (Calc6)
	Calc6Anterior: number;
	Calc6Atual: number;
	Calc6TesteAtual: string;
	Calc6ADV: number;

	// 🔥 NOVOS CAMPOS (Calc7)
	Calc7Anterior: number;
	Calc7Atual: number;
	Calc7TesteAtual: string;
	Calc7ADV: number;

	BaseMinimaAtual: string;
	BaseMinimaAnterior: string;
}